# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Surplus API Documentation",
        "description": "L5 Swagger OpenApi Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "get list of products",
                "description": "products list",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "create product",
                "description": "create product",
                "operationId": "createproduct",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "image_ids",
                                    "category_ids"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "product name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "description",
                                        "example": "test desc"
                                    },
                                    "image_ids": {
                                        "description": "image_ids",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "format": "id",
                                            "example": 1
                                        }
                                    },
                                    "category_ids": {
                                        "description": "category_ids",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "format": "id",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully Created"
                    },
                    "422": {
                        "description": "Unprocessable Content - Validation"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "get product by id",
                "description": "get product by id",
                "operationId": "getproductbyid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "products"
                ],
                "summary": "update product by id",
                "description": "update product by id",
                "operationId": "updateproduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "image_ids",
                                    "category_ids"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "product name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "description",
                                        "example": "test desc"
                                    },
                                    "image_ids": {
                                        "description": "image_ids",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "format": "id",
                                            "example": 1
                                        }
                                    },
                                    "category_ids": {
                                        "description": "category_ids",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "format": "id",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Unprocessable Content - Validation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "delete product by id",
                "description": "delete product by Id",
                "operationId": "deleteproduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "delete product by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success - No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    }
}